+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/preferences/ProjectBuildConfigurationComponent.java
@@ -24,6 +24,7 @@ import melnorme.lang.ide.core.operations.build.BuildTargetData;
 import melnorme.lang.ide.core.operations.build.CommandInvocation;
 import melnorme.lang.ide.core.operations.build.ProjectBuildInfo;
 import melnorme.lang.ide.core.utils.ProjectValidator;
+import melnorme.lang.ide.ui.build.BuildTargetEditor;
 import melnorme.lang.ide.ui.launch.BuildTargetField;
 import melnorme.lang.ide.ui.preferences.common.IPreferencesEditor;
 import melnorme.lang.ide.ui.utils.UIOperationsStatusHandler;