+++ documentation/ChangeLog.md
@@ -8,6 +8,7 @@ It's now possible to specify a command other than the default one (the $$SDK too
  * Added a setting to invoke a special build target when a $IDE editor is saved.
    * This allows invoking a build command (informally called a "check-build") that only checks for compiler errors, but doesn't not produce binaries. This has the potential to be faster than a full build.
    * Default is `` for $$IDE
+ * Added support for modifying the environment variables of a Build Target's build command.  
   
  * Fixed "IllegalStateException: The service has been unregistered" on Mars.2 when Eclipse is closed.
  * Added signing to releases.